ingestion_pipeline_delphi/requirements.txt
requests==2.32.3
pandas==2.2.2
google-cloud-bigquery==3.25.0
pyarrow==16.1.0
dbt-bigquery==1.6.1
dbt-core==1.6.1